more work on animation importer

+++ TagTool/Commands/ModelAnimationGraphs/AddAnimationCommand.cs
@@ -128,7 +128,7 @@ namespace TagTool.Commands.ModelAnimationGraphs
                     CurrentCompression = ModelAnimationGraph.Animation.CompressionValue.BestAccuracy,
                     FrameCount = (short)importer.frameCount,
                     NodeCount = (sbyte)importer.AnimationNodes.Count,
-                    NodeListChecksum = importer.nodeChecksum,
+                    NodeListChecksum = (int)importer.nodeChecksum,
                     Unknown2 = 5, //don't know what these do, but set usual values
                     Unknown3 = 6,
                     Heading = new RealVector3d(1,0,0),
